+++ eras/allegra/impl/test/Main.hs
@@ -8,6 +8,8 @@ import qualified Test.Cardano.Ledger.Allegra.BinarySpec as BinarySpec
 import qualified Test.Cardano.Ledger.Allegra.Imp as Imp
 import Test.Cardano.Ledger.Allegra.ImpTest ()
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
+import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
 
 main :: IO ()
 main =
@@ -17,3 +19,5 @@ main =
       CddlSpec.spec
       describe "Imp" $ do
         Imp.spec @AllegraEra
+      roundTripJsonEraSpec @AllegraEra
+      roundTripJsonShelleyEraSpec @AllegraEra