+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Transition.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -16,16 +16,7 @@ import Cardano.Ledger.BaseTypes (toKeyValuePairs)
 import Cardano.Ledger.Mary
 import Cardano.Ledger.Mary.Transition (TransitionConfig (MaryTransitionConfig))
 import Cardano.Ledger.Shelley.Transition
-import Data.Aeson (
-  FromJSON (..),
-  KeyValue (..),
-  ToJSON (..),
-  Value (..),
-  object,
-  pairs,
-  withObject,
-  (.:),
- )
+import Data.Aeson (KeyValue (..))
 import GHC.Generics
 import Lens.Micro
 import NoThunks.Class (NoThunks (..))
@@ -49,21 +40,6 @@ instance EraTransition AlonzoEra where
 
 instance NoThunks (TransitionConfig AlonzoEra)
 
-instance ToJSON (TransitionConfig AlonzoEra) where
-  toJSON = object . toAlonzoTransitionConfigPairs
-  toEncoding = pairs . mconcat . toAlonzoTransitionConfigPairs
-
 toAlonzoTransitionConfigPairs :: KeyValue e a => TransitionConfig AlonzoEra -> [a]
-toAlonzoTransitionConfigPairs alonzoConfig =
-  toShelleyTransitionConfigPairs shelleyConfig
-    ++ ["alonzo" .= object (toKeyValuePairs (alonzoConfig ^. tcTranslationContextL))]
-  where
-    maryConfig = alonzoConfig ^. tcPreviousEraConfigL
-    allegraConfig = maryConfig ^. tcPreviousEraConfigL
-    shelleyConfig = allegraConfig ^. tcPreviousEraConfigL
-
-instance FromJSON (TransitionConfig AlonzoEra) where
-  parseJSON = withObject "AlonzoTransitionConfig" $ \o -> do
-    pc <- parseJSON (Object o)
-    ag <- o .: "alonzo"
-    pure $ mkTransitionConfig pc ag
+toAlonzoTransitionConfigPairs = toKeyValuePairs
+{-# DEPRECATED toAlonzoTransitionConfigPairs "In favor of `toKeyValuePairs`" #-}